refactor: LoginUserArgumentResolver의 검증 책임을 AuthService로 이동 및 Repository 조회 로직 제거
