+++ be/src/main/java/com/jjikmuk/sikdorak/user/auth/api/LoginUserArgumentResolver.java
@@ -1,0 +1,35 @@
+package com.jjikmuk.sikdorak.user.auth.api;
+
+import com.jjikmuk.sikdorak.user.auth.app.AuthService;
+import java.util.Objects;
+import javax.servlet.http.HttpServletRequest;
+import lombok.RequiredArgsConstructor;
+import org.springframework.core.MethodParameter;
+import org.springframework.stereotype.Component;
+import org.springframework.web.bind.support.WebDataBinderFactory;
+import org.springframework.web.context.request.NativeWebRequest;
+import org.springframework.web.method.support.HandlerMethodArgumentResolver;
+import org.springframework.web.method.support.ModelAndViewContainer;
+
+@Component
+@RequiredArgsConstructor
+public class LoginUserArgumentResolver implements HandlerMethodArgumentResolver {
+
+
+	private final AuthService authService;
+
+	@Override
+	public boolean supportsParameter(MethodParameter parameter) {
+		return parameter.hasParameterAnnotation(AuthenticatedUser.class);
+	}
+
+	@Override
+	public Object resolveArgument(MethodParameter parameter, ModelAndViewContainer mavContainer,
+		NativeWebRequest webRequest, WebDataBinderFactory binderFactory) {
+
+		HttpServletRequest httpServletRequest = Objects.requireNonNull(
+			webRequest.getNativeRequest(HttpServletRequest.class));
+
+		return authService.authenticate(httpServletRequest);
+	}
+}
