+++ be/src/main/java/com/jjikmuk/sikdorak/user/auth/app/AuthService.java
@@ -1,8 +1,12 @@
 package com.jjikmuk.sikdorak.user.auth.app;
 
+import com.jjikmuk.sikdorak.user.auth.api.LoginUser;
 import com.jjikmuk.sikdorak.user.auth.app.dto.JwtTokenPair;
+import com.jjikmuk.sikdorak.user.auth.exception.InvalidTokenException;
 import com.jjikmuk.sikdorak.user.user.command.app.UserService;
 import com.jjikmuk.sikdorak.user.user.exception.NotFoundUserException;
+import java.util.Objects;
+import javax.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -11,6 +15,8 @@ import org.springframework.transaction.annotation.Transactional;
 @RequiredArgsConstructor
 public class AuthService {
 
+    private static final String AUTHORIZATION_HEADER = "Authorization";
+    private static final String TOKEN_TYPE = "Bearer ";
     private final UserService userService;
     private final JwtProvider jwtProvider;
 
@@ -25,6 +31,28 @@ public class AuthService {
         }
 
         return jwtProvider.createTokenResponse(userId);
+    }
+
+    public LoginUser authenticate(HttpServletRequest request) {
+
+        String token;
+        try {
+            token = parseAuthorizationHeader(request);
+            jwtProvider.validateToken(token);
+        }  catch (InvalidTokenException e) {
+            return LoginUser.anonymous();
+        }
+
+        Long userId = Long.valueOf(jwtProvider.decodeToken(token));
+        return LoginUser.user(userId);
+    }
+
+    private String parseAuthorizationHeader(HttpServletRequest request) {
+        String authorization = request.getHeader(AUTHORIZATION_HEADER);
+        if (Objects.isNull(authorization) || authorization.isEmpty()) {
+            throw new InvalidTokenException();
+        }
+        return authorization.replace(TOKEN_TYPE, "");
     }
 
     //TODO: RefreshToken을 레디스에 저장하는 로직 만들기
